Add more QuestionSettings stories
